backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pyodbc==5.1.0
pymssql==2.3.0
python-dotenv==1.0.1
google-generativeai==0.7.2
passlib==1.7.4
bcrypt==4.1.3
python-jose==3.3.0
python-multipart==0.0.9
pydantic==2.7.1
